requirements.txt
amqp==1.4.9
anyjson==0.3.3
bcrypt==3.1.3
billiard==3.3.0.23
celery==3.1.18
cffi==1.10.0
Django==1.11.3
ffmpy==0.2.2
kombu==3.0.37
olefile==0.44
Pillow==4.2.1
pycparser==2.18
pytz==2017.2
six==1.10.0
vine==1.1.4
